Minimum sample sizes for the threshold analysis

This object was produced by finding the point at which df_sim crossed the target power threshold of 80% (see details).

data(df_ss)

Format

A data.frame of 6840 rows and 15 columns. The first 14 columns give parameter combinations that were used in simulating and analysing data. The final "N_opt" column gives the optimal sample size to achieve a power of 80%.

Details

Minimum sample sizes were calculated as follows:

Find the value of N that crosses the threshold, and the value of N preceding it that does not.

Do linear interpolation between these two values to get the estimated sample size at the threshold.

Deal with special cases of N always being below the target power or always above the target power.

Some additional manual wrangling of final results. Ensure that N always decreases with increasing numbers of clusters (this is not always the case due to random variation).

Examples

data(df_ss)