Get pre-computed sample size tables

Text

get_sample_size_table(ICC = 0.05)

Arguments

ICC

the assumed intra-cluster correlation. Allowed values are 1, 2.

Details

The function get_power() was run over a large range of parameter combinations and results were stored within the df_sim object (see ?df_sim). These simulations were then used to produce minimum sample size estimates by linear interpolation that were stored within the df_ss object (see ?df_ss). This function provides a simple way of querying the df_ss object for given parameter values.

Examples

get_sample_size_presence(n_clust = 5, prevalence = 0.01, ICC = 0.1)
#> [1] 294